fix: useBackNavigation
