+++ pages/community/leave/types/[slug].tsx
@@ -1,9 +1,9 @@
 import { NextPage } from "next";
 import { useRouter } from "next/router";
+import { useEffect } from "react";
 
 import ContentLayout from "~community/common/components/templates/ContentLayout/ContentLayout";
 import ROUTES from "~community/common/constants/routes";
-import { useRouteChangeHandler } from "~community/common/hooks/useRouteChangeHandler";
 import { useTranslator } from "~community/common/hooks/useTranslator";
 import LeaveTypeActivationToggleButton from "~community/leave/components/molecules/LeaveTypeActivationToggleButton/LeaveTypeActivationToggleButton";
 import UnsavedChangesModal from "~community/leave/components/molecules/UserPromptModals/UnsavedChangesModal/UnsavedChangesModal";
@@ -40,16 +40,34 @@ const LeaveType: NextPage = () => {
       stopAllOngoingQuickSetup: state.stopAllOngoingQuickSetup
     }));
 
-  useRouteChangeHandler({
-    preventNavigation: () => isLeaveTypeFormDirty && !isLeaveTypeModalOpen,
-    onBeforeRouteAbort: (url) => {
-      setPendingNavigation(url);
-      setLeaveTypeModalType(LeaveTypeModalEnums.UNSAVED_CHANGES_MODAL);
-    },
-    onRouteChange: () => {
-      resetEditingLeaveType();
-    }
-  });
+  useEffect(() => {
+    const handleRouteChangeStart = (url: string) => {
+      if (ongoingQuickSetup.SETUP_LEAVE_TYPES) {
+        stopAllOngoingQuickSetup();
+      }
+      if (isLeaveTypeFormDirty && !isLeaveTypeModalOpen) {
+        setPendingNavigation(url);
+        setLeaveTypeModalType(LeaveTypeModalEnums.UNSAVED_CHANGES_MODAL);
+        router.events.emit("routeChangeError");
+        throw "routeChange aborted";
+      } else {
+        resetEditingLeaveType();
+      }
+    };
+
+    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
+      event.preventDefault();
+      return "";
+    };
+
+    router.events.on("routeChangeStart", handleRouteChangeStart);
+    window.addEventListener("beforeunload", handleBeforeUnload);
+
+    return () => {
+      router.events.off("routeChangeStart", handleRouteChangeStart);
+      window.removeEventListener("beforeunload", handleBeforeUnload);
+    };
+  }, [ongoingQuickSetup, isLeaveTypeFormDirty, isLeaveTypeModalOpen]);
 
   const handleBackBtnClick = () => {
     if (ongoingQuickSetup.SETUP_LEAVE_TYPES) {
